implement new contact card in contact preview + refactor useCase for jobROle+contactName
