--- packages/apps/frontera/src/routes/organization/src/components/Tabs/panels/PeoplePanel/components/ContactCard/components/index.ts
@@ -1,2 +1,0 @@
-export * from './ContactCardMenu';
-export * from './ContactLocation';
