+++ packages/apps/frontera/src/routes/src/components/ContactDetails/index.ts
@@ -1,0 +1,1 @@
+export * from './ContactDetails';
